backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
cycler==0.10.0
defusedxml==0.6.0
dj-rest-auth==2.1.3
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.6.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.0
patsy==0.5.1
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.6.0
seaborn==0.10.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
statsmodels==0.12.1
urllib3==1.26.2
XlsxWriter==1.3.7
